+++ src/test/kotlin/lotto/model/LottoDrawingMachineTest.kt
